Make setup.sh more flexible
Warn the user first if their not running this script as root.

Bump Ubuntu version to 18.04.1.

Prompt the user if the expected Linux distribution is not installed.

Use three variables to control which categories of applications to
install and setup: LLVM, other dev tools, and theme/aesthetics.

Add some extra echo newlines for clarity.